+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## Unreleased
+
+__Added__
+
+* Filenames can be provided via STDIN when not provided on the command line.
+
 ## 0.1.0 (2018/05/24)
 
 __Added__